рефакт пакетов с тестовыми классами
