+++ hw03-reflection/src/test/java/my/alkarps/engine/EngineTest.java
@@ -5,7 +5,8 @@ import my.alkarps.engine.exception.execute.InvokeSupportMethodException;
 import my.alkarps.engine.exception.validate.ClassNotFoundException;
 import my.alkarps.engine.exception.validate.*;
 import my.alkarps.engine.helper.ResultCaptor;
-import my.alkarps.engine.helper.notvalid.*;
+import my.alkarps.engine.helper.fail.*;
+import my.alkarps.engine.helper.invalid.*;
 import my.alkarps.engine.helper.valid.*;
 import my.alkarps.engine.model.ClassDetails;
 import my.alkarps.engine.model.Statistics;
