+++ hw03-reflection/src/test/java/my/alkarps/engine/helper/fail/TestClassWithTestMethodAndFailAfterEachMethod.java
@@ -1,4 +1,4 @@
-package my.alkarps.engine.helper.valid;
+package my.alkarps.engine.helper.fail;
 
 import my.alkarps.annotation.AfterEach;
 import my.alkarps.annotation.BeforeEach;
