+++ hw03-reflection/src/test/java/my/alkarps/engine/helper/invalid/TestClassWithOnlyAfterAllMethods.java
@@ -1,6 +1,6 @@
-package my.alkarps.engine.helper.notvalid;
+package my.alkarps.engine.helper.invalid;
 
-import my.alkarps.annotation.AfterEach;
+import my.alkarps.annotation.AfterAll;
 
 import static org.junit.jupiter.api.Assertions.fail;
 
@@ -8,24 +8,24 @@ import static org.junit.jupiter.api.Assertions.fail;
  * @author alkarps
  * create date 17.07.2020 18:15
  */
-public class TestClassWithOnlyAfterEachMethods {
-    @AfterEach
-    public void setDown() {
+public class TestClassWithOnlyAfterAllMethods {
+    @AfterAll
+    public static void setDown() {
         fail("Вызван метод, хотя не должен");
     }
 
-    @AfterEach
-    protected void setDown2() {
+    @AfterAll
+    protected static void setDown2() {
         fail("Вызван метод, хотя не должен");
     }
 
-    @AfterEach
-    private void setDown3() {
+    @AfterAll
+    private static void setDown3() {
         fail("Вызван метод, хотя не должен");
     }
 
-    @AfterEach
-    void setDown4() {
+    @AfterAll
+    static void setDown4() {
         fail("Вызван метод, хотя не должен");
     }
 }
