+++ hw03-reflection/src/test/java/my/alkarps/engine/helper/invalid/TestClassWithOnlyBeforeAllMethods.java
@@ -1,28 +1,31 @@
-package my.alkarps.engine.helper.notvalid;
+package my.alkarps.engine.helper.invalid;
+
+import my.alkarps.annotation.BeforeAll;
 
 import static org.junit.jupiter.api.Assertions.fail;
 
 /**
  * @author alkarps
- * create date 17.07.2020 17:48
+ * create date 17.07.2020 18:13
  */
-public class TestClassWithPublicConstructorButNotTestMethods {
-    public TestClassWithPublicConstructorButNotTestMethods() {
-    }
-
-    public void publicMethod(){
+public class TestClassWithOnlyBeforeAllMethods {
+    @BeforeAll
+    public static void setUp() {
         fail("Вызван метод, хотя не должен");
     }
 
-    private void privateMethod(){
+    @BeforeAll
+    protected static void setUp2() {
         fail("Вызван метод, хотя не должен");
     }
 
-    protected void protectedMethod(){
+    @BeforeAll
+    private static void setUp3() {
         fail("Вызван метод, хотя не должен");
     }
 
-    void packageMethod(){
+    @BeforeAll
+    static void setUp4() {
         fail("Вызван метод, хотя не должен");
     }
 }
